// src/interfaces/salida-ai.interface.ts
export interface SalidaGeminiInterface{    

    rateOmision: number

    rateComision: number

    rateExactitud: number

    puntajeCoherencia: number
 
    puntajeFluidez: number

    puntajeTotal: number

    detallesOmitidos:string[]

    palabrasClaveOmitidas: string[]

    aciertos: string[]

    conclusion: string
}